remove semgrep.yml and fix error
by having semgrep.yml as an empty file as I decided to use semgrep-rules to make it easier for me to understand and not have two files name the same. Semgrep ended up scanning the empty file and causing an error during the deployment. fixing error and cleaning up code.

W edit